Split the Settings component into multiple files
